requirements.txt
ffmpeg-python
python-ffmpeg-video-streaming
azure-storage-blob